+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Better mapping of styles in DOCX to HTML / Markdown converter
 - Option to handle custom font families as code inline (preformatted text) in DOCX to Markdown converter
 - Support emphasis and images inside hyperlinks in DOCX to Markdown converter
+- Support base64 images in Markdown to DOCX/RTF converter
 - Fix: unnecessary space was created in DOCX to Markdown converter in some cases
 - Fix other minor bugs and improve code structure